Dockerize the app
For making the development environment more similiar to the production
environment docker is used. Two containers were created, one for
Django+gunicorn and the other one for nginx.

For development with the Django dev server one can use the
'docker-compose.dev.yaml' file. While using
'docker-compose.dev_full-stack.yaml' will run the Django app using
gunicorn and nginx - still a development setup, but it should be close
to a production setup.

The production setup is still missing and needs to be setup in the
future prior to roll out.